# [v2.10] Faster agent image build