--- a/Makefile
+++ b/Makefile
@@ -17,7 +17,13 @@ $(TARGETS): .dapper
 
 .DEFAULT_GOAL := ci
 
+quick-agent:
+	@$(MAKE) quick TARGET="agent"
+
+quick-server:
+	@$(MAKE) quick TARGET="server"
+
 $(DEV_TARGETS):
 	./dev-scripts/$@
 
-.PHONY: $(TARGETS) $(DEV_TARGETS)
+.PHONY: $(TARGETS) $(DEV_TARGETS) quick-agent quick-server