--- a/dev-scripts/quick
+++ b/dev-scripts/quick
@@ -31,32 +31,36 @@ xargs -n2 docker save -o ./k3s-airgap-images.tar <<< "${AIRGAP_IMAGES}"
 # download kontainer driver metadata
 curl -sLf https://releases.rancher.com/kontainer-driver-metadata/"${CATTLE_KDM_BRANCH}"/data.json > ./data.json
 
-# start the builds
-docker buildx build \
-  --build-arg VERSION="${TAG}" \
-  --build-arg ARCH="${ARCH}" \
-  --build-arg IMAGE_REPO="${REPO}" \
-  --build-arg COMMIT="${COMMIT}" \
-  --build-arg RKE_VERSION="${RKE_VERSION}" \
-  --build-arg CATTLE_RANCHER_WEBHOOK_VERSION="${CATTLE_RANCHER_WEBHOOK_VERSION}" \
-  --build-arg CATTLE_RANCHER_PROVISIONING_CAPI_VERSION="${CATTLE_RANCHER_PROVISIONING_CAPI_VERSION}" \
-  --build-arg CATTLE_CSP_ADAPTER_MIN_VERSION="${CATTLE_CSP_ADAPTER_MIN_VERSION}" \
-  --build-arg CATTLE_FLEET_VERSION="${CATTLE_FLEET_VERSION}" \
-  --tag "${REPO}"/rancher:"${TAG}" \
-  --platform="${OS}/${ARCH}" \
-  --target server \
-  --file ./package/Dockerfile .
+if [ -z $TARGET ] || [ $TARGET = "server" ]; then
+  # start the builds
+  docker buildx build \
+    --build-arg VERSION="${TAG}" \
+    --build-arg ARCH="${ARCH}" \
+    --build-arg IMAGE_REPO="${REPO}" \
+    --build-arg COMMIT="${COMMIT}" \
+    --build-arg RKE_VERSION="${RKE_VERSION}" \
+    --build-arg CATTLE_RANCHER_WEBHOOK_VERSION="${CATTLE_RANCHER_WEBHOOK_VERSION}" \
+    --build-arg CATTLE_RANCHER_PROVISIONING_CAPI_VERSION="${CATTLE_RANCHER_PROVISIONING_CAPI_VERSION}" \
+    --build-arg CATTLE_CSP_ADAPTER_MIN_VERSION="${CATTLE_CSP_ADAPTER_MIN_VERSION}" \
+    --build-arg CATTLE_FLEET_VERSION="${CATTLE_FLEET_VERSION}" \
+    --tag "${REPO}"/rancher:"${TAG}" \
+    --platform="${OS}/${ARCH}" \
+    --target server \
+    --file ./package/Dockerfile .
+fi
 
-docker buildx build \
-  --build-arg VERSION="${TAG}" \
-  --build-arg ARCH="${ARCH}" \
-  --build-arg RANCHER_TAG="${TAG}" \
-  --build-arg RANCHER_REPO="${REPO}" \
-  --build-arg COMMIT="${COMMIT}" \
-  --build-arg RKE_VERSION="${RKE_VERSION}" \
-  --build-arg CATTLE_RANCHER_WEBHOOK_VERSION="${CATTLE_RANCHER_WEBHOOK_VERSION}" \
-  --build-arg CATTLE_RANCHER_PROVISIONING_CAPI_VERSION="${CATTLE_RANCHER_PROVISIONING_CAPI_VERSION}" \
-  --tag "${REPO}"/rancher-agent:"${TAG}" \
-  --platform="${OS}/${ARCH}" \
-  --target agent \
-  --file ./package/Dockerfile .
+if [ -z $TARGET ] || [ $TARGET = "agent" ]; then
+  docker buildx build \
+    --build-arg VERSION="${TAG}" \
+    --build-arg ARCH="${ARCH}" \
+    --build-arg RANCHER_TAG="${TAG}" \
+    --build-arg RANCHER_REPO="${REPO}" \
+    --build-arg COMMIT="${COMMIT}" \
+    --build-arg RKE_VERSION="${RKE_VERSION}" \
+    --build-arg CATTLE_RANCHER_WEBHOOK_VERSION="${CATTLE_RANCHER_WEBHOOK_VERSION}" \
+    --build-arg CATTLE_RANCHER_PROVISIONING_CAPI_VERSION="${CATTLE_RANCHER_PROVISIONING_CAPI_VERSION}" \
+    --tag "${REPO}"/rancher-agent:"${TAG}" \
+    --platform="${OS}/${ARCH}" \
+    --target agent \
+    --file ./package/Dockerfile .
+fi