--- a/scripts/package
+++ b/scripts/package
@@ -32,6 +32,7 @@ docker build \
     --build-arg CATTLE_RANCHER_PROVISIONING_CAPI_VERSION="${CATTLE_RANCHER_PROVISIONING_CAPI_VERSION}" \
     --build-arg CATTLE_CSP_ADAPTER_MIN_VERSION="${CATTLE_CSP_ADAPTER_MIN_VERSION}" \
     --build-arg CATTLE_FLEET_VERSION="${CATTLE_FLEET_VERSION}" \
+    --target server \
     -t ${IMAGE} .
 
 docker build \
@@ -41,7 +42,8 @@ docker build \
     --build-arg CATTLE_RANCHER_WEBHOOK_VERSION="${CATTLE_RANCHER_WEBHOOK_VERSION}" \
     --build-arg CATTLE_RANCHER_PROVISIONING_CAPI_VERSION="${CATTLE_RANCHER_PROVISIONING_CAPI_VERSION}" \
     --build-arg RANCHER_REPO=${REPO} \
-    -t ${AGENT_IMAGE} -f Dockerfile.agent .
+    --target agent \
+    -t ${AGENT_IMAGE} .
 
 mkdir -p ../dist
 echo ${IMAGE} > ../dist/images