--- a/tests/v2/codecoverage/scripts/build_docker_images.sh
+++ b/tests/v2/codecoverage/scripts/build_docker_images.sh
@@ -32,6 +32,7 @@ docker build \
   --build-arg CATTLE_RANCHER_PROVISIONING_CAPI_VERSION="${CATTLE_RANCHER_PROVISIONING_CAPI_VERSION}" \
   --build-arg CATTLE_CSP_ADAPTER_MIN_VERSION="${CATTLE_CSP_ADAPTER_MIN_VERSION}" \
   --build-arg CATTLE_FLEET_VERSION="${CATTLE_FLEET_VERSION}" \
+  --target server \
   -t ${IMAGE} -f Dockerfile . --no-cache
 
 echo "building agent test docker image"
@@ -42,7 +43,8 @@ docker build \
   --build-arg RANCHER_REPO=${REPO} \
   --build-arg CATTLE_RANCHER_WEBHOOK_VERSION="${CATTLE_RANCHER_WEBHOOK_VERSION}" \
   --build-arg CATTLE_RANCHER_PROVISIONING_CAPI_VERSION="${CATTLE_RANCHER_PROVISIONING_CAPI_VERSION}" \
-  -t ${AGENT_IMAGE} -f Dockerfile.agent . --no-cache
+  --target agent \
+  -t ${AGENT_IMAGE} -f Dockerfile . --no-cache
 
 echo ${DOCKERHUB_PASSWORD} | docker login --username ${DOCKERHUB_USERNAME} --password-stdin