Backport of https://github.com/rancher/rancher/pull/49636 and https://github.com/rancher/rancher/pull/49963. Description below is taken from original image + tweaked.

2.10 is slightly different because it includes system-charts and rancher/charts from the master branch. So did the same here where they end up on their own intermediate stages.

---

## Issue

We have `package/Dockerfile` for the Rancher server and `package/Dockerfile.agent` for the agent. Because of this, it's difficult to share stages between the two final image. This currently means that in order to build the agent, we first have to completely build the Rancher image. However, looking closer, we only need a few things from the final Rancher image.

There's a few ways to avoid building all of Rancher before being able to build just the agent:
1. Use a single Dockerfile and specify which `--target` when building. **This is the approach of this PR.**
2. Duplicate the steps in both files. That's more maintenance to keep in sync. Not a fan.
3. Have an intermediate build target that can be used to build both rancher and agent. Building the agent would require two docker buildx command.

First, the results..

## Results

With the improvements from https://github.com/rancher/rancher/pull/49633 is now much quicker for a edit-compile workflow (from ~3m30s to ~30s).

Now if we take this same edit-compile workflow but we're interested in only building one image, then we get the following result:

Result (when editing a single file)

||Before|After|
|:-- | ---:| ---:|
|Building only rancher|29s|13s|
|Building only agent|29s|13s|

## Explaining the changes

**Use multi-stage build**

`tini`, `rancher-charts`, `partner-charts`, `rke2-charts` and `kdm` are now downloaded in their own intermediate stage. This removes dependencies between them and the rest of the variables and makes it easier to share those steps between the server and the agent build. Another improvement here is that those are done in parallel now.

**chroot-builder and friends**

I made sure to keep the same installed packages from the old images. So we get a common `chroot-builder` stage for common dependency and a `chroot-builder-<target>` for packages that must only be in one or the other. This way, resulting images are exactly the same as they were before. (We can decide to change which packages should be part of which images but let's do that in another PR)

**Added `make quick-agent` and `make quick-server` Makefile targets**

An easy to remember (and works with bash autocomplete) way to build either the agent or server image only. So now these commands are available: `make quick` (builds both), `make quick-agent` (builds only the agent) and `make quick-server` (builds only the server).

**Parallel CI steps for building server and agent**

The old CI's `build-agent` job depended on the server image to be built, which meant we had to build the server first in a job, move that to an artifact, and then build the agent (pulling from the artifact). Now that the agent no longer depends on the server image, we can remove this dependency.

~~I'm not sure why we're not building both image in the same job, it sounds like that would be much quicker because they would be able to leverage Docker's caching feature. Now instead they both need to pull the bci images, install the dependencies, run go mod download, etc.~~ Tested it and it's not as fast as building them in parallel, need more investigation here..

## Resulting image

The resulting image ends up being pretty much the same. Here's the script I'm using to look at the image size, image content and env vars defined (**Note:** had to change kdm branch and chart branch to dev-v2.11 for this test because it was pointing to release-v2.11 since we're doing the release)

```sh
#!/bin/sh

tag1="v2.10-894572416f01"
tag2="single-dockerfile-v2.10"

docker run -it --entrypoint sh docker.io/rancher/rancher:$tag1 -c 'find / | grep -v "^/proc" | sort' > server-find-$tag1.txt
docker run -it --entrypoint sh docker.io/rancher/rancher:$tag2 -c 'find / | grep -v "^/proc" | sort' > server-find-$tag2.txt

echo "Comparing server image files"
diff -u server-find-$tag1.txt server-find-$tag2.txt

docker run -it --entrypoint sh docker.io/rancher/rancher-agent:$tag1 -c 'find / | grep -v "^/proc" |sort' > agent-find-$tag1.txt
docker run -it --entrypoint sh docker.io/rancher/rancher-agent:$tag2 -c 'find / | grep -v "^/proc" |sort' > agent-find-$tag2.txt

echo "Comparing agent image files"
diff -u agent-find-$tag1.txt agent-find-$tag2.txt

docker run -it --entrypoint env docker.io/rancher/rancher:$tag1 > server-env-$tag1.txt
docker run -it --entrypoint env docker.io/rancher/rancher:$tag2 > server-env-$tag2.txt

echo "Comparing server image env vars"
diff -u server-env-$tag1.txt server-env-$tag2.txt

docker run -it --entrypoint env docker.io/rancher/rancher-agent:$tag1 > agent-env-$tag1.txt
docker run -it --entrypoint env docker.io/rancher/rancher-agent:$tag2 > agent-env-$tag2.txt

echo "Comparing agent image env vars"
diff -u agent-env-$tag1.txt agent-env-$tag2.txt

echo "Comparing server image size"
docker inspect -f "{{ .Size }}" docker.io/rancher/rancher:$tag1
docker inspect -f "{{ .Size }}" docker.io/rancher/rancher:$tag2

echo "Comparing agent image size"
docker inspect -f "{{ .Size }}" docker.io/rancher/rancher-agent:$tag1
docker inspect -f "{{ .Size }}" docker.io/rancher/rancher-agent:$tag2
```

Here are the results:

```
Comparing server image size
2574016373
2456172404
Comparing agent image size
1566755196
1450240534
```

<details>
<summary>Server image content comparison</summary>

```diff
--- server-find-v2.10-894572416f01.txt  2025-04-24 15:08:34.906861661 -0400                                                                                                                                                                                                                                            
+++ server-find-single-dockerfile-v2.10.txt     2025-04-24 15:08:36.163534181 -0400                                                                                                                                                                                                                                    
@@ -68801,6 +68801,7 @@                                                                                                                                                                                                                                                                                                
 /usr/share/help/zh_TW                                                                                                                                                                                                                                                                                                 
 /usr/share/icons                                                                                                                                                                                                                                                                                                      
 /usr/share/info                                                                                                                                                                                                                                                                                                       
+/usr/share/info/dir                                                                                                                                                                                                                                                                                                   
 /usr/share/java                                                                                                                                                                                                                                                                                                       
 /usr/share/licenses                                                                                                                                                                                                                                                                                                   
 /usr/share/licenses/bash                                                                                                                                                                                                                                                                                              
@@ -71801,9 +71802,9 @@                                                                                                                                                                                                                                                                                                
 /var/lib/rancher-data/local-catalogs/library/.git/objects                                                                                                                                                                                                                                                             
 /var/lib/rancher-data/local-catalogs/library/.git/objects/info                                                                                                                                                                                                                                                        
 /var/lib/rancher-data/local-catalogs/library/.git/objects/pack                                                                                                                                                                                                                                                        
-/var/lib/rancher-data/local-catalogs/library/.git/objects/pack/pack-d416ebc4c679c4a5af36a12e3c38a83bb7b22a3e.idx                                                                                                                                                                                                      
-/var/lib/rancher-data/local-catalogs/library/.git/objects/pack/pack-d416ebc4c679c4a5af36a12e3c38a83bb7b22a3e.pack                                                                                                                                                                                                     
-/var/lib/rancher-data/local-catalogs/library/.git/objects/pack/pack-d416ebc4c679c4a5af36a12e3c38a83bb7b22a3e.rev                                                                                                                                                                                                      
+/var/lib/rancher-data/local-catalogs/library/.git/objects/pack/pack-2bf2e92fec5586d6f1690daaabb484ad65b4a6f0.idx                                                                                                                                                                                                      
+/var/lib/rancher-data/local-catalogs/library/.git/objects/pack/pack-2bf2e92fec5586d6f1690daaabb484ad65b4a6f0.pack                                                                                                                                                                                                     
+/var/lib/rancher-data/local-catalogs/library/.git/objects/pack/pack-2bf2e92fec5586d6f1690daaabb484ad65b4a6f0.rev                                                                                                                                                                                                      
 /var/lib/rancher-data/local-catalogs/library/.git/packed-refs                                                                                                                                                                                                                                                         
 /var/lib/rancher-data/local-catalogs/library/.git/refs                                                                                                                                                                                                                                                                
 /var/lib/rancher-data/local-catalogs/library/.git/refs/heads                                                                                                                                                                                                                                                          
@@ -120218,9 +120219,9 @@                                                                                                                                                                                                                                                                                              
 /var/lib/rancher-data/local-catalogs/v2/rancher-rke2-charts/675f1b63a0a83905972dcab2794479ed599a6f41b86cd6193d69472d0fa889c9/.git/objects                                                                                                                                                                             
 /var/lib/rancher-data/local-catalogs/v2/rancher-rke2-charts/675f1b63a0a83905972dcab2794479ed599a6f41b86cd6193d69472d0fa889c9/.git/objects/info                                                                                                                                                                        
 /var/lib/rancher-data/local-catalogs/v2/rancher-rke2-charts/675f1b63a0a83905972dcab2794479ed599a6f41b86cd6193d69472d0fa889c9/.git/objects/pack                                                                                                                                                                        
-/var/lib/rancher-data/local-catalogs/v2/rancher-rke2-charts/675f1b63a0a83905972dcab2794479ed599a6f41b86cd6193d69472d0fa889c9/.git/objects/pack/pack-d7606a5cf640f32c095a4a74ba4514427f6b07ff.idx                                                                                                                      
-/var/lib/rancher-data/local-catalogs/v2/rancher-rke2-charts/675f1b63a0a83905972dcab2794479ed599a6f41b86cd6193d69472d0fa889c9/.git/objects/pack/pack-d7606a5cf640f32c095a4a74ba4514427f6b07ff.pack                                                                                                                     
-/var/lib/rancher-data/local-catalogs/v2/rancher-rke2-charts/675f1b63a0a83905972dcab2794479ed599a6f41b86cd6193d69472d0fa889c9/.git/objects/pack/pack-d7606a5cf640f32c095a4a74ba4514427f6b07ff.rev                                                                                                                      
+/var/lib/rancher-data/local-catalogs/v2/rancher-rke2-charts/675f1b63a0a83905972dcab2794479ed599a6f41b86cd6193d69472d0fa889c9/.git/objects/pack/pack-74b9f2d9c1febc8915945a77093f481f9dc95387.idx                                                                                                                      
+/var/lib/rancher-data/local-catalogs/v2/rancher-rke2-charts/675f1b63a0a83905972dcab2794479ed599a6f41b86cd6193d69472d0fa889c9/.git/objects/pack/pack-74b9f2d9c1febc8915945a77093f481f9dc95387.pack                                                                                                                     
+/var/lib/rancher-data/local-catalogs/v2/rancher-rke2-charts/675f1b63a0a83905972dcab2794479ed599a6f41b86cd6193d69472d0fa889c9/.git/objects/pack/pack-74b9f2d9c1febc8915945a77093f481f9dc95387.rev                                                                                                                      
 /var/lib/rancher-data/local-catalogs/v2/rancher-rke2-charts/675f1b63a0a83905972dcab2794479ed599a6f41b86cd6193d69472d0fa889c9/.git/packed-refs                                                                                                                                                                         
 /var/lib/rancher-data/local-catalogs/v2/rancher-rke2-charts/675f1b63a0a83905972dcab2794479ed599a6f41b86cd6193d69472d0fa889c9/.git/refs                                                                                                                                                                                
 /var/lib/rancher-data/local-catalogs/v2/rancher-rke2-charts/675f1b63a0a83905972dcab2794479ed599a6f41b86cd6193d69472d0fa889c9/.git/refs/heads
```
</details>

<details>
<summary>Agent image content comparison</summary>

```diff
--- agent-find-v2.10-894572416f01.txt   2025-04-24 15:08:37.150205443 -0400                                                                                                                                                                                                                                            
+++ agent-find-single-dockerfile-v2.10.txt      2025-04-24 15:08:38.280210705 -0400                                                                                                                                                                                                                                    
@@ -70138,6 +70138,7 @@                                                                                                                                                                                                                                                                                                
 /usr/share/help/zh_TW                                                                                                                                                                                                                                                                                                 
 /usr/share/icons                                                                                                                                                                                                                                                                                                      
 /usr/share/info                                                                                                                                                                                                                                                                                                       
+/usr/share/info/dir                                                                                                                                                                                                                                                                                                   
 /usr/share/java
 /usr/share/kbd
 /usr/share/kbd/consolefonts
@@ -73447,9 +73448,9 @@
 /var/lib/rancher-data/local-catalogs/library/.git/objects
 /var/lib/rancher-data/local-catalogs/library/.git/objects/info
 /var/lib/rancher-data/local-catalogs/library/.git/objects/pack
-/var/lib/rancher-data/local-catalogs/library/.git/objects/pack/pack-d416ebc4c679c4a5af36a12e3c38a83bb7b22a3e.idx
-/var/lib/rancher-data/local-catalogs/library/.git/objects/pack/pack-d416ebc4c679c4a5af36a12e3c38a83bb7b22a3e.pack
-/var/lib/rancher-data/local-catalogs/library/.git/objects/pack/pack-d416ebc4c679c4a5af36a12e3c38a83bb7b22a3e.rev
+/var/lib/rancher-data/local-catalogs/library/.git/objects/pack/pack-2bf2e92fec5586d6f1690daaabb484ad65b4a6f0.idx
+/var/lib/rancher-data/local-catalogs/library/.git/objects/pack/pack-2bf2e92fec5586d6f1690daaabb484ad65b4a6f0.pack
…